fixing direct page visits

+++ delete/index.php
@@ -1,15 +1,29 @@
 <?php
-  // Access the database
+  // Proceed only if form data has been received
+  if (!isset($_POST['form_name'])) {
+    // unset($_SESSION['form_name']);
+    header('Location: ./../view');
+    exit();
+  }
+
+  // Retrieve database connection code
   $assetPointer = "../assets";
   require "$assetPointer/dbConnect.php";
+  
+  // Establish database connection
   $db = get_db($assetPointer);
   
+  // Prepare the DELETE statement
   $statement = $db->prepare("DELETE FROM trip WHERE driver_id = 1;");  
+  // Execute the statement
   $statement->execute();
+  /*
   if (file_exists('../record/data.json')) {
     unlink('../record/data.json');
   }
-  header('Location: ../view');
+  */
+  header('Location: ./../view');
+  exit();
 ?>
 
 <!DOCTYPE html>
